requirements.txt
pycryptodome
aiohttp
slixmpp